// front-end/src/modules/user/Variations/variations.types.ts
export interface IVariationsListDetails {
  attachment_id: string;
  company_id: number;
  contract_id: number;
  contract_name: string;
  id: string;
  project_id: number;
  project_name: string;
  variation_amount: number;
  variation_id: number;
  variation_name: string;
  variation_status: string;
  created_on: string;
  file_name: string;
  file_path: string;
}